CI: fix the three failures it found
Exactly what CI is for — none of these could be seen on Windows.

twk_receive takes unsigned long long*, which is not uint64_t on LP64 platforms
even though it is the same width. MSVC never noticed because there they are the
same type; gcc and clang both rejected it, so every Linux and macOS job failed
on the JNI shim.

kCCSuccess lives in CommonCryptoError.h, which the other CommonCrypto headers
only sometimes pull in — the macOS crypto backend did not compile.

The C# harness publishes to bin/x64/Release, not bin/Release, so the smoke step
copied twk.dll to a path that did not exist. It now publishes to a fixed
directory rather than guessing the SDK's layout.

Worth recording what already passed: Linux built and tested green with the
OpenSSL backend and the libcurl host — both written blind and never compiled
until now — including the crypto known-answer tests.

+++ bindings/android/src/main/cpp/twk_jni.cpp
@@ -295,8 +295,8 @@ JNIEXPORT void JNICALL Java_org_ton_walletkit_core_Native_send(JNIEnv* env, jcla
                                                                jbyteArray params) {
     const std::string method_utf8 = from_bytes(env, method);
     const std::string params_utf8 = from_bytes(env, params);
-    twk_send(reinterpret_cast<twk_client*>(client), static_cast<uint64_t>(request_id), method_utf8.c_str(),
-             params_utf8.c_str());
+    twk_send(reinterpret_cast<twk_client*>(client), static_cast<unsigned long long>(request_id),
+             method_utf8.c_str(), params_utf8.c_str());
 }
 
 /**
@@ -306,7 +306,9 @@ JNIEXPORT void JNICALL Java_org_ton_walletkit_core_Native_send(JNIEnv* env, jcla
 JNIEXPORT jbyteArray JNICALL Java_org_ton_walletkit_core_Native_receive(JNIEnv* env, jclass, jlong client,
                                                                        jdouble timeout,
                                                                        jlongArray request_id_out) {
-    uint64_t request_id = 0;
+    // The ABI says unsigned long long, which is not uint64_t on LP64 platforms
+    // even though it is the same width — so the declaration follows the header.
+    unsigned long long request_id = 0;
     const char* message = twk_receive(reinterpret_cast<twk_client*>(client), timeout, &request_id);
     if (message == nullptr) {
         return nullptr;
